MAINT: Add injection script for true-fullscreen mac
This patch adds an injection script to inject fullscreen into a loaded
application to allow for displaying the app above the notch on a
macbook. -- uses `frida` to inject.

Some changes to nvim-tree have been made to ensure we can navigate from
relative path instead of defaulting to the project root always.

CHANGES:

- Adds injection script for displayOverNotch
- Adds the respective key-bind for skhd.
- Updates to nvim-tree behaviour