MAINT: Small change to support workspaces structure in work projects
This patch just moves to my fork of `neoscopes` and ensures that we
select the root workspace on startup to allow us to not receive errors
in `neovim` when using `find files`.

Have submitted a pull request to the repository.

TODO: Remember to move back to the original repo if the PR gets accepted.