MAINT: Add FZF searching instead of regex match parsing to searches
This allows for easier searching and restricting scope of the search, as
we can now match on greps using the filepath as part of the match.

i.e: an fzf match `filesyscom . !spec . messageService`, will match for
all paths referencing `file-system.component` that are not spec files,
and then some inner file text that includes `messagingService`.

Where `.` denotes the AND operator, `|` is the OR operand for FZF, more
information can be found in the man pages.