GH-1: Extend functionality of harpoon to include tmux list
The initial patch for this (this patch), is to cut down the original
implementation in `harpoon/master`. Further simplifications can be made,
but the goal will be to rewrite some of this to query windows instead of
controlling them with `go_to_terminal`

(i.e make harpoon use the windows that the session has instead of relying
on keeping the indexes in memory and referencing them) [NOTE: don't know
if this is a good idea yet, its small enough that we can just implement
it and test it out].

This first patch is a basic implementation that only supports a single
terminal to send commands to.

Other notes: a couple of random dotfiles changes were included in this
patch because I am lazy