MAINT: Prevent halting of terminal on start after installing `nvm`
NVM Utility was taking a while to startup, implemented a lazy loading
function to allow for loading the utility once and then reusing the util
once it is loaded.

This allows us to only do the bootstrap once when the command is first
run, instead of doing it when the .zprofile is loaded (when a terminal
is started it hangs).

NOTE: This can probably be prevented by loading it in another location
that exports the paths before getting into the `zsh` shell, this
solution seems fine for now though.